attempt at making class for each run
