+++ idrstream_cp_env.yml
@@ -4,25 +4,25 @@ channels:
 - conda-forge
 
 dependencies:
-- python=3.8.13
-- jupyter=1.0.0
-- pandas=1.4.2
-- pytorch=1.11.0
-- scikit-image=0.19.3
-- pyimagej=1.2.1
-- pip=22.1.2
-- numpy
-- matplotlib
-- pandas
-- openjdk
-- scikit-learn
-- mahotas
-- gtk2
-- Jinja2=3.0.1
-- inflect=5.3.0
-- wxpython=4.1.0
-- mysqlclient=1.4.4
-- sentry-sdk=0.18.0
+- conda-forge::python=3.8.13
+- conda-forge::jupyter=1.0.0
+- conda-forge::pandas=1.4.2
+- conda-forge::pytorch=1.11.0
+- conda-forge::scikit-image=0.19.3
+- conda-forge::pyimagej=1.2.1
+- conda-forge::pip=22.1.2
+- conda-forge::numpy
+- conda-forge::matplotlib
+- conda-forge::pandas
+- conda-forge::openjdk
+- conda-forge::scikit-learn
+- conda-forge::mahotas
+- conda-forge::gtk2
+- conda-forge::Jinja2=3.0.1
+- conda-forge::inflect=5.3.0
+- conda-forge::wxpython=4.1.0
+- conda-forge::mysqlclient=1.4.4
+- conda-forge::sentry-sdk=0.18.0
 - pip:
   - cellprofiler==4.2.4
   - cellpose==2.1.0
